requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
flask-login
gunicorn
flask-login